remove op genesis conversion from reth-chainspec
