+++ crates/chainspec/src/spec.rs
@@ -590,15 +590,7 @@ impl ChainSpec {
 
 impl From<Genesis> for ChainSpec {
     fn from(genesis: Genesis) -> Self {
-        #[cfg(not(feature = "optimism"))]
-        {
-            into_ethereum_chain_spec(genesis)
-        }
-
-        #[cfg(feature = "optimism")]
-        {
-            into_optimism_chain_spec(genesis)
-        }
+        into_ethereum_chain_spec(genesis)
     }
 }
 
@@ -638,7 +630,6 @@ impl EthereumHardforks for ChainSpec {
 impl reth_optimism_forks::OptimismHardforks for ChainSpec {}
 
 /// Convert the given [`Genesis`] into an Ethereum [`ChainSpec`].
-#[cfg(not(feature = "optimism"))]
 fn into_ethereum_chain_spec(genesis: Genesis) -> ChainSpec {
     // Block-based hardforks
     let hardfork_opts = [
@@ -722,105 +713,6 @@ fn into_ethereum_chain_spec(genesis: Genesis) -> ChainSpec {
         hardforks: ChainHardforks::new(ordered_hardforks),
         paris_block_and_final_difficulty,
         deposit_contract,
-        ..Default::default()
-    }
-}
-
-#[cfg(feature = "optimism")]
-/// Convert the given [`Genesis`] into an Optimism [`ChainSpec`].
-fn into_optimism_chain_spec(genesis: Genesis) -> ChainSpec {
-    use reth_optimism_forks::OptimismHardfork;
-    let optimism_genesis_info = OptimismGenesisInfo::extract_from(&genesis);
-    let genesis_info = optimism_genesis_info.optimism_chain_info.genesis_info.unwrap_or_default();
-
-    // Block-based hardforks
-    let hardfork_opts = [
-        (EthereumHardfork::Homestead.boxed(), genesis.config.homestead_block),
-        (EthereumHardfork::Tangerine.boxed(), genesis.config.eip150_block),
-        (EthereumHardfork::SpuriousDragon.boxed(), genesis.config.eip155_block),
-        (EthereumHardfork::Byzantium.boxed(), genesis.config.byzantium_block),
-        (EthereumHardfork::Constantinople.boxed(), genesis.config.constantinople_block),
-        (EthereumHardfork::Petersburg.boxed(), genesis.config.petersburg_block),
-        (EthereumHardfork::Istanbul.boxed(), genesis.config.istanbul_block),
-        (EthereumHardfork::MuirGlacier.boxed(), genesis.config.muir_glacier_block),
-        (EthereumHardfork::Berlin.boxed(), genesis.config.berlin_block),
-        (EthereumHardfork::London.boxed(), genesis.config.london_block),
-        (EthereumHardfork::ArrowGlacier.boxed(), genesis.config.arrow_glacier_block),
-        (EthereumHardfork::GrayGlacier.boxed(), genesis.config.gray_glacier_block),
-        (OptimismHardfork::Bedrock.boxed(), genesis_info.bedrock_block),
-    ];
-    let mut block_hardforks = hardfork_opts
-        .into_iter()
-        .filter_map(|(hardfork, opt)| opt.map(|block| (hardfork, ForkCondition::Block(block))))
-        .collect::<Vec<_>>();
-
-    // Paris
-    let paris_block_and_final_difficulty =
-        if let Some(ttd) = genesis.config.terminal_total_difficulty {
-            block_hardforks.push((
-                EthereumHardfork::Paris.boxed(),
-                ForkCondition::TTD {
-                    total_difficulty: ttd,
-                    fork_block: genesis.config.merge_netsplit_block,
-                },
-            ));
-
-            genesis.config.merge_netsplit_block.map(|block| (block, ttd))
-        } else {
-            None
-        };
-
-    // Time-based hardforks
-    let time_hardfork_opts = [
-        (EthereumHardfork::Shanghai.boxed(), genesis.config.shanghai_time),
-        (EthereumHardfork::Cancun.boxed(), genesis.config.cancun_time),
-        (EthereumHardfork::Prague.boxed(), genesis.config.prague_time),
-        (OptimismHardfork::Regolith.boxed(), genesis_info.regolith_time),
-        (OptimismHardfork::Canyon.boxed(), genesis_info.canyon_time),
-        (OptimismHardfork::Ecotone.boxed(), genesis_info.ecotone_time),
-        (OptimismHardfork::Fjord.boxed(), genesis_info.fjord_time),
-        (OptimismHardfork::Granite.boxed(), genesis_info.granite_time),
-    ];
-
-    let mut time_hardforks = time_hardfork_opts
-        .into_iter()
-        .filter_map(|(hardfork, opt)| opt.map(|time| (hardfork, ForkCondition::Timestamp(time))))
-        .collect::<Vec<_>>();
-
-    block_hardforks.append(&mut time_hardforks);
-
-    // Ordered Hardforks
-    let mainnet_hardforks = OptimismHardfork::op_mainnet();
-    let mainnet_order = mainnet_hardforks.forks_iter();
-
-    let mut ordered_hardforks = Vec::with_capacity(block_hardforks.len());
-    for (hardfork, _) in mainnet_order {
-        if let Some(pos) = block_hardforks.iter().position(|(e, _)| **e == *hardfork) {
-            ordered_hardforks.push(block_hardforks.remove(pos));
-        }
-    }
-
-    // append the remaining unknown hardforks to ensure we don't filter any out
-    ordered_hardforks.append(&mut block_hardforks);
-
-    // NOTE: in full node, we prune all receipts except the deposit contract's. We do not
-    // have the deployment block in the genesis file, so we use block zero. We use the same
-    // deposit topic as the mainnet contract if we have the deposit contract address in the
-    // genesis json.
-    let deposit_contract = genesis.config.deposit_contract_address.map(|address| DepositContract {
-        address,
-        block: 0,
-        topic: MAINNET_DEPOSIT_CONTRACT.topic,
-    });
-
-    ChainSpec {
-        chain: genesis.config.chain_id.into(),
-        genesis,
-        genesis_hash: OnceCell::new(),
-        hardforks: ChainHardforks::new(ordered_hardforks),
-        paris_block_and_final_difficulty,
-        deposit_contract,
-        base_fee_params: optimism_genesis_info.base_fee_params,
         ..Default::default()
     }
 }
@@ -1099,59 +991,6 @@ impl DepositContract {
     /// Creates a new [`DepositContract`].
     pub const fn new(address: Address, block: BlockNumber, topic: B256) -> Self {
         Self { address, block, topic }
-    }
-}
-
-/// Genesis info for Optimism.
-#[cfg(feature = "optimism")]
-#[derive(Default, Debug, serde::Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct OptimismGenesisInfo {
-    optimism_chain_info: op_alloy_rpc_types::genesis::OptimismChainInfo,
-    #[serde(skip)]
-    base_fee_params: BaseFeeParamsKind,
-}
-
-#[cfg(feature = "optimism")]
-impl OptimismGenesisInfo {
-    fn extract_from(genesis: &Genesis) -> Self {
-        let mut info = Self {
-            optimism_chain_info: op_alloy_rpc_types::genesis::OptimismChainInfo::extract_from(
-                &genesis.config.extra_fields,
-            )
-            .unwrap_or_default(),
-            ..Default::default()
-        };
-        if let Some(optimism_base_fee_info) = &info.optimism_chain_info.base_fee_info {
-            if let (Some(elasticity), Some(denominator)) = (
-                optimism_base_fee_info.eip1559_elasticity,
-                optimism_base_fee_info.eip1559_denominator,
-            ) {
-                let base_fee_params = if let Some(canyon_denominator) =
-                    optimism_base_fee_info.eip1559_denominator_canyon
-                {
-                    BaseFeeParamsKind::Variable(
-                        vec![
-                            (
-                                EthereumHardfork::London.boxed(),
-                                BaseFeeParams::new(denominator as u128, elasticity as u128),
-                            ),
-                            (
-                                reth_optimism_forks::OptimismHardfork::Canyon.boxed(),
-                                BaseFeeParams::new(canyon_denominator as u128, elasticity as u128),
-                            ),
-                        ]
-                        .into(),
-                    )
-                } else {
-                    BaseFeeParams::new(denominator as u128, elasticity as u128).into()
-                };
-
-                info.base_fee_params = base_fee_params;
-            }
-        }
-
-        info
     }
 }
 
@@ -2477,7 +2316,6 @@ Post-merge hard forks (timestamp based):
     }
 
     #[test]
-    #[cfg(not(feature = "optimism"))]
     fn test_fork_order_ethereum_mainnet() {
         let genesis = Genesis {
             config: ChainConfig {
@@ -2526,82 +2364,6 @@ Post-merge hard forks (timestamp based):
             EthereumHardfork::Paris.boxed(),
             EthereumHardfork::Shanghai.boxed(),
             EthereumHardfork::Cancun.boxed(),
-        ];
-
-        assert!(expected_hardforks
-            .iter()
-            .zip(hardforks.iter())
-            .all(|(expected, actual)| &**expected == *actual));
-        assert_eq!(expected_hardforks.len(), hardforks.len());
-    }
-
-    #[test]
-    #[cfg(feature = "optimism")]
-    fn test_fork_order_optimism_mainnet() {
-        use reth_optimism_forks::OptimismHardfork;
-
-        let genesis = Genesis {
-            config: ChainConfig {
-                chain_id: 0,
-                homestead_block: Some(0),
-                dao_fork_block: Some(0),
-                dao_fork_support: false,
-                eip150_block: Some(0),
-                eip155_block: Some(0),
-                eip158_block: Some(0),
-                byzantium_block: Some(0),
-                constantinople_block: Some(0),
-                petersburg_block: Some(0),
-                istanbul_block: Some(0),
-                muir_glacier_block: Some(0),
-                berlin_block: Some(0),
-                london_block: Some(0),
-                arrow_glacier_block: Some(0),
-                gray_glacier_block: Some(0),
-                merge_netsplit_block: Some(0),
-                shanghai_time: Some(0),
-                cancun_time: Some(0),
-                terminal_total_difficulty: Some(U256::ZERO),
-                extra_fields: [
-                    (String::from("bedrockBlock"), 0.into()),
-                    (String::from("regolithTime"), 0.into()),
-                    (String::from("canyonTime"), 0.into()),
-                    (String::from("ecotoneTime"), 0.into()),
-                    (String::from("fjordTime"), 0.into()),
-                    (String::from("graniteTime"), 0.into()),
-                ]
-                .into_iter()
-                .collect(),
-                ..Default::default()
-            },
-            ..Default::default()
-        };
-
-        let chain_spec: ChainSpec = into_optimism_chain_spec(genesis);
-
-        let hardforks: Vec<_> = chain_spec.hardforks.forks_iter().map(|(h, _)| h).collect();
-        let expected_hardforks = vec![
-            EthereumHardfork::Homestead.boxed(),
-            EthereumHardfork::Tangerine.boxed(),
-            EthereumHardfork::SpuriousDragon.boxed(),
-            EthereumHardfork::Byzantium.boxed(),
-            EthereumHardfork::Constantinople.boxed(),
-            EthereumHardfork::Petersburg.boxed(),
-            EthereumHardfork::Istanbul.boxed(),
-            EthereumHardfork::MuirGlacier.boxed(),
-            EthereumHardfork::Berlin.boxed(),
-            EthereumHardfork::London.boxed(),
-            EthereumHardfork::ArrowGlacier.boxed(),
-            EthereumHardfork::GrayGlacier.boxed(),
-            EthereumHardfork::Paris.boxed(),
-            OptimismHardfork::Bedrock.boxed(),
-            OptimismHardfork::Regolith.boxed(),
-            EthereumHardfork::Shanghai.boxed(),
-            OptimismHardfork::Canyon.boxed(),
-            EthereumHardfork::Cancun.boxed(),
-            OptimismHardfork::Ecotone.boxed(),
-            OptimismHardfork::Fjord.boxed(),
-            OptimismHardfork::Granite.boxed(),
         ];
 
         assert!(expected_hardforks
